# Replace pow(2,x) by ldexp(1.0,x) in HyperLogLog for performance