@rui314 2023-07-20
Thanks. I don't know how much it will translate to the real-world performance difference, but we at least shouldn't lose anything with this change. I'll merge this change once all tests pass on CI.